+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Security   to invite users to upgrade in case of vulnerabilities.
 -->
 
 ## [Unreleased]
+### Added
+- New class `VcsVersionInfo` that wraps and enriches `VersionInfo` with the latest tag.
 
 ## [0.3.2] - 2016-03-16
 ### Added