+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Security   to invite users to upgrade in case of vulnerabilities.
 ## [Unreleased]
 ### Added
 - New class `VcsVersionInfo` that wraps and enriches `VersionInfo` with the latest tag.
+- New class `OS` with method `isWindows`.
 
 ## [0.3.2] - 2016-03-16
 ### Added