+++ CHANGELOG.md
@@ -18,6 +18,12 @@ Security   to invite users to upgrade in case of vulnerabilities.
 - New class `VcsVersionInfo` that wraps and enriches `VersionInfo` with the latest tag.
 - New class `OS` with methods `isWindows` and `hasBinary`.
 - New `Folder` methods `delete` and `deleteContents`
+- New `Strings` methods:
+    - `getCommonPrefix`
+    - `getCommonSuffix`
+    - `groupByCommonPrefix`
+    - `groupByCommonSuffix`
+    - `reverse`
 
 ## [0.3.2] - 2016-03-16
 ### Added